+++ taxcalc/tests/test_utils.py
@@ -11,9 +11,7 @@ import os
 import math
 import random
 import numpy as np
-from numpy.testing import assert_equal, assert_almost_equal, assert_array_equal
 import pandas as pd
-from pandas.util.testing import assert_series_equal
 import pytest
 # pylint: disable=import-error
 from taxcalc import Policy, Records, Behavior, Calculator, ParametersBase
@@ -62,7 +60,7 @@ def test_expand_1d_short_array():
     exp[3:] = exp2
     res = ParametersBase._expand_1D(ary, inflate=True,
                                     inflation_rates=[0.02] * 10, num_years=10)
-    assert np.allclose(exp, res, atol=0.0, rtol=1.0E-7)
+    assert np.allclose(exp, res, atol=0.01, rtol=0.0)
 
 
 def test_expand_1d_variable_rates():
@@ -71,7 +69,8 @@ def test_expand_1d_variable_rates():
     res = ParametersBase._expand_1D(ary, inflate=True,
                                     inflation_rates=irates, num_years=5)
     exp = np.array([4, 5, 9, 9 * 1.03, 9 * 1.03 * 1.035])
-    assert np.allclose(exp.astype('f4', casting='unsafe'), res)
+    assert np.allclose(exp.astype('f4', casting='unsafe'), res,
+                       atol=0.01, rtol=0.0)
 
 
 def test_expand_1d_scalar():
@@ -79,7 +78,7 @@ def test_expand_1d_scalar():
     exp = np.array([val * math.pow(1.02, i) for i in range(0, 10)])
     res = ParametersBase._expand_1D(val, inflate=True,
                                     inflation_rates=[0.02] * 10, num_years=10)
-    assert np.allclose(exp, res)
+    assert np.allclose(exp, res, atol=0.01, rtol=0.0)
 
 
 def test_expand_1d_accept_none():
@@ -95,7 +94,8 @@ def test_expand_1d_accept_none():
     exp = np.array(exp)
     res = ParametersBase._expand_array(lst, inflate=True,
                                        inflation_rates=irates, num_years=5)
-    assert np.allclose(exp.astype('f4', casting='unsafe'), res)
+    assert np.allclose(exp.astype('f4', casting='unsafe'), res,
+                       atol=0.01, rtol=0.0)
 
 
 def test_expand_2d_short_array():
@@ -108,7 +108,7 @@ def test_expand_2d_short_array():
     exp[1:] = exp2
     res = ParametersBase._expand_2D(ary, inflate=True,
                                     inflation_rates=[0.02] * 5, num_years=5)
-    assert np.allclose(exp, res)
+    assert np.allclose(exp, res, atol=0.01, rtol=0.0)
 
 
 def test_expand_2d_variable_rates():
@@ -127,7 +127,7 @@ def test_expand_2d_variable_rates():
     exp[1:] = exp2
     res = ParametersBase._expand_2D(ary, inflate=True,
                                     inflation_rates=irates, num_years=5)
-    assert np.allclose(exp, res)
+    assert np.allclose(exp, res, atol=0.01, rtol=0.0)
 
 
 def test_validity_of_name_lists():
@@ -188,13 +188,13 @@ def test_weighted_count_lt_zero():
     diffs = grped.apply(weighted_count_lt_zero, 'tax_diff')
     exp = pd.Series(data=[4, 0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
     df2 = pd.DataFrame(data=DATA_FLOAT, columns=['tax_diff', 's006', 'label'])
     grped = df2.groupby('label')
     diffs = grped.apply(weighted_count_lt_zero, 'tax_diff')
     exp = pd.Series(data=[4, 0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_weighted_count_gt_zero():
@@ -203,13 +203,13 @@ def test_weighted_count_gt_zero():
     diffs = grped.apply(weighted_count_gt_zero, 'tax_diff')
     exp = pd.Series(data=[8, 10], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
     df2 = pd.DataFrame(data=DATA, columns=['tax_diff', 's006', 'label'])
     grped = df2.groupby('label')
     diffs = grped.apply(weighted_count_gt_zero, 'tax_diff')
     exp = pd.Series(data=[8, 10], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_weighted_count():
@@ -218,7 +218,7 @@ def test_weighted_count():
     diffs = grouped.apply(weighted_count)
     exp = pd.Series(data=[12, 10], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_weighted_mean():
@@ -227,7 +227,7 @@ def test_weighted_mean():
     diffs = grouped.apply(weighted_mean, 'tax_diff')
     exp = pd.Series(data=[16.0 / 12.0, 26.0 / 10.0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_wage_weighted():
@@ -255,7 +255,7 @@ def test_weighted_sum():
     diffs = grouped.apply(weighted_sum, 'tax_diff')
     exp = pd.Series(data=[16.0, 26.0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_weighted_perc_inc():
@@ -264,7 +264,7 @@ def test_weighted_perc_inc():
     diffs = grouped.apply(weighted_perc_inc, 'tax_diff')
     exp = pd.Series(data=[8. / 12., 1.0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 def test_weighted_perc_dec():
@@ -273,7 +273,7 @@ def test_weighted_perc_dec():
     diffs = grouped.apply(weighted_perc_dec, 'tax_diff')
     exp = pd.Series(data=[4. / 12., 0.0], index=['a', 'b'])
     exp.index.name = 'label'
-    assert_series_equal(exp, diffs)
+    pd.util.testing.assert_series_equal(exp, diffs)
 
 
 EPSILON = 1e-5
@@ -406,9 +406,10 @@ def test_diff_table_sum_row(cps_subsample):
                       'aftertax_perc']
     digit_cols = [x for x in tdiff1.columns.tolist() if
                   x not in non_digit_cols]
-    assert np.allclose(tdiff1[digit_cols][-1:], tdiff2[digit_cols][-1:])
-    assert_array_equal(tdiff1[non_digit_cols][-1:],
-                       tdiff2[non_digit_cols][-1:])
+    assert np.allclose(tdiff1[digit_cols][-1:],
+                       tdiff2[digit_cols][-1:])
+    np.testing.assert_array_equal(tdiff1[non_digit_cols][-1:],
+                                  tdiff2[non_digit_cols][-1:])
 
 
 def test_row_classifier(cps_subsample):
@@ -432,7 +433,7 @@ def test_row_classifier(cps_subsample):
                                            result_type='weighted_sum',
                                            baseline_obj=calc1.records).s006
     # use weighted sum of weights in each cell to check classifer
-    assert_array_equal(calc1_s006, calc2_s006)
+    assert np.allclose(calc1_s006, calc2_s006, atol=0.00, rtol=0.0)
 
 
 def test_expand_2d_already_filled():
@@ -442,7 +443,7 @@ def test_expand_2d_already_filled():
                 [40000, 74900, 37450, 50200, 74900, 37450]]
     res = ParametersBase._expand_2D(_II_brk2, inflate=True,
                                     inflation_rates=[0.02] * 5, num_years=3)
-    assert_equal(res, np.array(_II_brk2))
+    np.allclose(res, np.array(_II_brk2), atol=0.01, rtol=0.0)
 
 
 def test_expand_2d_partial_expand():
@@ -466,7 +467,7 @@ def test_expand_2d_partial_expand():
            [exp1, exp2, exp3, exp4, exp5, exp6]]
     res = ParametersBase._expand_2D(_II_brk2, inflate=True,
                                     inflation_rates=inf_rates, num_years=4)
-    assert_equal(res, exp)
+    assert np.allclose(res, exp, atol=0.01, rtol=0.0)
 
 
 def test_strip_nones():
@@ -495,7 +496,7 @@ def test_expand_2d_accept_none():
     exp = np.array(exp).astype('i4', casting='unsafe')
     res = ParametersBase._expand_array(_II_brk2, inflate=True,
                                        inflation_rates=[0.02] * 5, num_years=4)
-    assert_equal(res, exp)
+    assert np.allclose(res, exp, atol=0.01, rtol=0.0)
 
     syr = 2013
     pol = Policy(start_year=syr)
@@ -508,7 +509,7 @@ def test_expand_2d_accept_none():
     exp_2019 = [41000.] + [(1.0 + irates[2018 - syr]) * i
                            for i in _II_brk2[2][1:]]
     exp_2019 = np.array(exp_2019)
-    assert_equal(pol.II_brk2, exp_2019)
+    assert np.allclose(pol.II_brk2, exp_2019, atol=0.01, rtol=0.0)
 
 
 def test_expand_2d_accept_none_add_row():
@@ -537,7 +538,7 @@ def test_expand_2d_accept_none_add_row():
     res = ParametersBase._expand_array(_II_brk2, inflate=True,
                                        inflation_rates=inflation_rates,
                                        num_years=5)
-    assert_equal(res, exp)
+    assert np.allclose(res, exp, atol=0.01, rtol=0.0)
     user_mods = {2016: {'_II_brk2': _II_brk2}}
     syr = 2013
     pol = Policy(start_year=syr)
@@ -551,7 +552,7 @@ def test_expand_2d_accept_none_add_row():
                            (1 + irates[2018 - syr]) * i
                            for i in _II_brk2[2][1:]]
     exp_2020 = np.array(exp_2020)
-    assert np.allclose(pol.II_brk2, exp_2020)
+    assert np.allclose(pol.II_brk2, exp_2020, atol=0.01, rtol=0.0)
 
 
 def test_mtr_graph_data(cps_subsample):
@@ -688,8 +689,8 @@ def test_myr_diag_table_wo_behv(cps_subsample):
     adt = multiyear_diagnostic_table(calc, 1)
     # extract combined liabilities as a float and
     # adopt units of the raw calculator data in liabilities_x
-    liabilities_y = adt.iloc[19].tolist()[0] * 1000000000
-    assert_almost_equal(liabilities_x, liabilities_y, 2)
+    liabilities_y = adt.iloc[19].tolist()[0] * 1e9
+    assert np.allclose(liabilities_x, liabilities_y, atol=0.01, rtol=0.0)
 
 
 def test_myr_diag_table_w_behv(cps_subsample):
@@ -712,7 +713,7 @@ def test_myr_diag_table_w_behv(cps_subsample):
     # extract combined liabilities as a float and
     # adopt units of the raw calculator data in liabilities_x
     liabilities_y = adt.iloc[19].tolist()[0] * 1e9
-    assert_almost_equal(liabilities_x, liabilities_y, 2)
+    assert np.allclose(liabilities_x, liabilities_y, atol=0.01, rtol=0.0)
 
 
 def test_ce_aftertax_income(cps_subsample):