+++ RELEASES.md
@@ -13,7 +13,15 @@ Release 0.9.3 on 2017-??-??
 - None
 
 **New Features**
-- None
+- Add validation checking of policy parameters involved in a reform
+  [[#1502](https://github.com/open-source-economics/Tax-Calculator/pull/1502)
+  by Martin Holmer]
+- Add option to use policy parameter suffixes in JSON reform files
+  [[#1505](https://github.com/open-source-economics/Tax-Calculator/pull/1505)
+  by Martin Holmer]
+- Add rounding of wage-inflated or price-inflated parameter values to nearest cent
+  [[#1506](https://github.com/open-source-economics/Tax-Calculator/pull/1506)
+  by Martin Holmer]
 
 **Bug Fixes**
 - None